refactor: Move callback to its proper location
looks like it got left in the wrong place during testing of multi-client
