+++ naff/models/naff/listener.py
@@ -2,7 +2,7 @@ import asyncio
 import inspect
 from typing import Coroutine, Callable
 
-from callback import CallbackObject
+from naff.models.naff.callback import CallbackObject
 from naff.api.events.internal import BaseEvent
 from naff.client.const import MISSING, Absent
 from naff.client.utils import get_event_name
